refactor: 컴포넌트 디렉터리 구조 변경
components 내부에 BaseComponent, common, lunch 3개의 상위 카테고리를 정한 후 각 카테고리에 맞는 컴포넌트만 추가하도록 변경
