--- src/components/SortDropdown/SortDropdown.type.ts
@@ -1,3 +1,0 @@
-export type SortCategory = "이름순" | "거리순";
-
-export type SortKey = "name" | "distance";
