+++ src/components/common/CommonDropdown/CommonDropdown.ts
@@ -1,11 +1,11 @@
-import BaseComponent from "../BaseComponent/BaseComponent";
-import type { CustomEventListenerDictionary } from "../BaseComponent/BaseComponent.type";
+import BaseComponent from "../../BaseComponent/BaseComponent";
+import type { CustomEventListenerDictionary } from "../../BaseComponent/BaseComponent.type";
 
-import { $ } from "../../utils/dom";
-import { createOptionElements } from "../../utils/createOptionElements";
-import { isCustomEventType } from "../../utils/typeGuard";
+import { $ } from "../../../utils/dom";
+import { createOptionElements } from "../../../utils/createOptionElements";
+import { isCustomEventType } from "../../../utils/typeGuard";
 
-import { CUSTOM_EVENT_TYPE } from "../../constants/eventType";
+import { CUSTOM_EVENT_TYPE } from "../../../constants/eventType";
 
 class CommonDropdown extends BaseComponent {
   private eventListeners: CustomEventListenerDictionary = {
