+++ src/domain/Restaurant/Restaurant.ts
@@ -1,12 +1,12 @@
 import RestaurantStorage from "../../storages/RestaurantStorage";
 import type { CustomStorage } from "../../storages/type";
 
-import type { SortCategory } from "../../components/SortDropdown/SortDropdown.type";
 import type { MenuCategory, RestaurantDetail } from "./Restaurant.type";
 
 import { MENU_CATEGORIES } from "../../constants/menuCategory/menuCategory";
 import { SORT_CATEGORIES_TYPE } from "../../constants/sortCategory/sortCategory";
 import { ERROR_MESSAGE } from "../../constants/errorMessage";
+import type { SortCategory } from "../../constants/sortCategory/sortCategory.type";
 
 class Restaurant {
   private currentCategory: MenuCategory = MENU_CATEGORIES.all;
